Delete actions.js
